requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-multipart>=0.0.6
email-validator>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.24.0

# GCP
google-cloud-storage>=2.8.0
google-cloud-pubsub>=2.13.11
google-cloud-functions>=1.13.0
google-auth>=2.16.2

# Data handling
redis>=4.5.4
aioredis>=2.0.1

# Video processing
ffmpeg-python>=0.2.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Utilities
tenacity>=8.2.2
python-dotenv>=1.0.0